+++ lib/components/SelectionControl/index.ts
@@ -1,1 +1,1 @@
-export * from "./checkbox";
+export * from "./checkboxs";
